+++ src/lang/lang.ml
@@ -782,8 +782,27 @@ let source_methods =
     ( "remaining",
       ([], fun_t [] float_t),
       "Estimation of remaining time in the current track.",
-      fun s -> val_fun [] (fun _ -> float (Frame.seconds_of_main s#remaining))
-    );
+      fun s ->
+        val_fun [] (fun _ ->
+            float
+              (let r = s#remaining in
+               if r < 0 then infinity else Frame.seconds_of_main r)) );
+    ( "elapsed",
+      ([], fun_t [] float_t),
+      "Elapsed time in the current track.",
+      fun s ->
+        val_fun [] (fun _ ->
+            float
+              (let e = s#elapsed in
+               if e < 0 then infinity else Frame.seconds_of_main e)) );
+    ( "duration",
+      ([], fun_t [] float_t),
+      "Estimation of the duration of the current track.",
+      fun s ->
+        val_fun [] (fun _ ->
+            float
+              (let d = s#duration in
+               if d < 0 then infinity else Frame.seconds_of_main d)) );
     ( "self_sync",
       ([], fun_t [] bool_t),
       "Is the source currently controling its own real-time loop.",